+++ carts/models.py
@@ -1,0 +1,18 @@
+from django.db import models
+from django.conf import settings
+from stores.models import Product
+
+
+class Cart(models.Model):
+    user = models.OneToOneField(settings.AUTH_USER_MODEL, on_delete=models.CASCADE, related_name='cart')
+    product = models.ForeignKey(Product, on_delete=models.CASCADE)
+    quantity = models.PositiveIntegerField()
+
+    def __str__(self):
+        return f"{self.user.username}'의 장바구니"
+
+    def get_total_price(self):
+        return self.product.price * self.quantity
+
+class CartItem(models.Model):
+    cart = models.ForeignKey(Cart, on_delete=models.CASCADE, related_name='cartitems')
