+++ carts/templates/carts/carts.html
@@ -1,0 +1,5 @@
+{% extends 'base.html' %}
+
+{% block content %}
+<h1>cart</h1>
+{% endblock content %}
