+++ carts/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+from . import views
+
+
+app_name = 'carts'
+
+urlpatterns = [
+    path('', views.cart, name='cart'),
+]
