+++ carts/views.py
@@ -1,0 +1,5 @@
+from django.shortcuts import render
+
+
+def cart(request):
+    return render(request, 'carts/carts.html')
