+++ planethelper/urls.py
@@ -26,13 +26,15 @@ from . import views
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('', views.main, name="main"),
+    path('accounts/', include('accounts.urls')),
+    path('accounts/', include('allauth.urls')),
+    path('posts/', include('posts.urls')),
     path('stores/', include('stores.urls')),
+    path('carts/', include('carts.urls')),
+
     path('ckeditor/', include('ckeditor_uploader.urls')),
     # path(r'^upload/', login_required(views_ckeditor.upload), name='ckeditor_upload'),
     # path(r'^browse/', never_cache(login_required(views_ckeditor.browse)), name='ckeditor_browse'),
     path('upload/', login_required(views_ckeditor.upload), name='ckeditor_upload'),
     path('browse/', never_cache(login_required(views_ckeditor.browse)), name='ckeditor_browse'),
-    path('posts/', include('posts.urls')),
-    path('accounts/', include('accounts.urls')),
-    path('accounts/', include('allauth.urls')),
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
