+++ stores/admin.py
@@ -6,6 +6,6 @@ admin.site.register(Store)
 admin.site.register(Product)
 admin.site.register(ProductReview)
 
-admin.site.register(Cart)
+# admin.site.register(Cart)
 admin.site.register(Order)
 admin.site.register(OrderItem)
