+++ templates/header.html
@@ -41,6 +41,9 @@
             <a class="nav-link active" href="{% url 'accounts:profile' request.user.username %}">프로필</a>
           </li>
           <li class="nav-item">
+            <a class="nav-link active" href="{% url 'carts:cart' %}">장바구니</a>
+          </li>
+          <li class="nav-item">
             <a class="nav-link active " href="{% url 'accounts:logout' %}">로그아웃</a>
           </li>
           {% else %}
